Basic Spawner - WIP
Spawner has an array of enemies it can spawn with max amounts of each type and probabilities of spawning.

Spawns only when the player is nearby and on a slightly randomised delay.

Spawns in a particle effect that then spawns in the character.

ISSUE - enemies spawned are not affected by gravity and do not function as they should